refactor(api-audit-logs): DI-native audit logs GraphQL schema factory
Replace the legacy createGraphQLSchemaPlugin + the generic registerResolvers bridge
(which wrapped each resolver as fn(parent,args,context,info), service-locating
context.auditLogs) with AuditLogsGraphQLSchema: a CoreGraphQLSchemaFactory using
builder.addResolver that declares the AuditLogsContext dependency (the facade holding
getAuditLog/listAuditLogs). Deletes graphql/schema.ts + the bridge helper from the
feature. Behavior-preserving; validation + resolve/resolveList unchanged.
